requirements.txt
Django>=2.1.5,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
django-cors-headers>=2.4.0,<2.5.0
djangorestframework-jwt>=1.11.0,<1.12.0

flake8>=3.6.0,<3.7.0